Перенёс коллбеки в родительский компонент
